requirement.txt
clip @ git+https://github.com/openai/CLIP.git
easydict
google-generativeai
openai
pycocoevalcap
word2number
